+++ index.html
@@ -11,6 +11,7 @@
 
 <body>
   <main>
+
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
       <h2 class="img-filters__title  visually-hidden">Фильтр фотографий</h2>
@@ -230,6 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src ="js/main.js"></script>
+  <script src="js/main.js " type="module"></script>
 </body>
 </html>
